A Bunch of Changes to the Main File
I updated this file in many ways. The screen region was changed to 800
by 600, to account for the border. The rest of the changes mainly had to
do with adding ways to switch between different states, and what
different menus those states would be. All of the state names are just
place holders, and are just used as a way to give an idea of what each
key would be doing. I added lots of comments explaining my thought
process and alternative methods that we may be able to use if the ones I
wrote are not optimal. I also added methods to use features that we may
or may not implement, such as different fire modes. I tried to organize
all the different controls by their type using comments. I'm sure there
is some of it that may run into problems, or is not optimal for our
game, so if you see anything that looks wrong or could be improved let
me know and I will change it.