An Edit to Enemy AI, and a program to change something's state.
Note the state changer cannot handle game, but will freeze the code at
that point when switching to other states, circumventing the need for a
timer.  It doesn't handle the changing of screen, but once the screen is
changed it can be used to hold the code in place until the state is
changed.