Lots of Bug Fixes by Dr. Galluzzi
This gets our files into a very odd, but working, state. It is missing a
lot and there is a lot that needs to be adjusted, but it works in this
state. It is important to note that you cannot run the files in the
folder called "3 Main Components," you need to run the IO (main) file
that is in the base ActionGameProject folder in order for it to load the
graphics.